+++ src/lightwin/core/em_fields/field100.py
@@ -38,17 +38,17 @@ class Field100(Field):
 
         Parameters
         ----------
-        path : pathlib.Path
+        path :
             The path to the ``EDZ`` file to load.
 
         Returns
         -------
-        e_z : Callable[[Pos1D], float]
+        e_z :
             Function that takes in ``z`` position and returns corresponding
             field, at null phase, for amplitude of :math:`1\,\mathrm{MV/m}`.
-        n_z : tuple[int]
+        n_z :
             Number of interpolation points.
-        n_cell : int
+        n_cell :
             Number of cell for cavities.
 
         """