+++ src/lightwin/optimisation/algorithms/differential_evolution.py
@@ -21,7 +21,6 @@ class DifferentialEvolution(DownhillSimplex):
 
         Returns
         -------
-        info : OptiSol
             Gives list of solutions, corresponding objective, convergence
             violation if applicable, etc.