+++ src/lightwin/optimisation/design_space/design_space.py
@@ -45,24 +45,19 @@ class DesignSpace:
 
         Parameters
         ----------
-        elements_names : Sequence[str]
+        elements_names :
             Name of the elements with variables and constraints.
-        filepath_variables : pathlib.Path
+        filepath_variables :
             Path to the ``variables.csv`` file.
-        variables_names : Sequence[str]
+        variables_names :
             Name of the variables to create.
-        filepath_constraints : pathlib.Path | None, optional
+        filepath_constraints :
             Path to the ``constraints.csv`` file. The default is None, in which
-            case no constraint is created.
-        constraints_names : Sequence[str] | None, optional
+        constraints_names :
             Name of the constraints to create. The default is None, in which
             case no constraint is created.
-        delimiter : str, optional
-            Delimiter in the files. The default is a commma.
-
-        Returns
-        -------
-        Self
+        delimiter :
+            Delimiter in the files.
 
         """
         variables = _from_file(
@@ -144,14 +139,14 @@ class DesignSpace:
 
         Parameters
         ----------
-        basepath : pathlib.Path
+        basepath :
             Folder where the files will be stored.
         variables_filename, constraints_filename : pathlib.Path | str, optional
             Name of the output files without extension.
-        overwrite : bool, optional
+        overwrite :
             To overwrite an existing file with the same name or not. The
             default is False.
-        to_csv_kw : dict[str, Any]
+        to_csv_kw :
             Keyword arguments given to the pandas ``to_csv`` method.
 
         """
@@ -192,11 +187,11 @@ class DesignSpace:
 
         Parameters
         ----------
-        parameters : list[DesignSpaceParameter]
+        parameters :
             All the defined parameters.
-        filepath : pathlib.Path
+        filepath :
             Where file will be stored.
-        delimiter : str
+        delimiter :
             Delimiter between two columns. The default is ','.
         to_csv_kw: dict[str, Any]
             Keyword arguments given to the pandas ``to_csv`` method.
@@ -219,15 +214,14 @@ class DesignSpace:
 
         Parameters
         ----------
-        element_name : str
+        element_name :
             Name of the element, which will be inserted in the output dict.
-        parameters : list[DesignSpaceParameter]
+        parameters :
             Parameters concerning the element, which ``limits`` (``x_0`` if
             appliable) will be inserted in the dict.
 
         Returns
         -------
-        dict[str, float | None | tuple[float, float]]
             Contains all :class:`.Variable` or :class:`.Constraint` information
             of the element.
 
@@ -265,14 +259,13 @@ def _gather_dicts_by_key(
 
     Parameters
     ----------
-    parameters : list[DesignSpaceParameter]
+    parameters :
         Objects to study.
-    key : str
+    key :
         Name of the attribute against which ``parameters`` should be gathered.
 
     Returns
     -------
-    dict[Any, list[DesignSpaceParameter]]
         Keys are all existing values of attribute ``key`` from ``parameters``.
         Values are lists of :class:`.DesignSpaceParameter` with ``key``
         attribute equaling the dict key.
@@ -295,14 +288,13 @@ def _parameters_to_dict(
 
     Parameters
     ----------
-    parameters : list[DesignSpaceParameter]
+    parameters :
         Where ``to_get`` will be looked for.
-    to_get : Sequence[str]
+    to_get :
         Values to get.
 
     Returns
     -------
-    list[dict]
         Contains ``to_get`` values in dictionaries for every parameter.
 
     """
@@ -352,20 +344,19 @@ def _from_file(
 
     Parameters
     ----------
-    parameter_class : type[Variable] | type[Constraint]
+    parameter_class :
         Object which ``from_pd_series`` method will be called.
-    filepath : pathlib.Path
+    filepath :
         Path to the ``CSV``.
-    elements_names : Sequence[str]
+    elements_names :
         Name of the elements.
-    parameters_names : Sequence[str]
+    parameters_names :
         Name of the parameters.
-    delimiter : str
+    delimiter :
         Delimiter in the ``CSV``.
 
     Returns
     -------
-    list[DesignSpaceParameter]
         List of variables or constraints.
 
     """