+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [0.12.x]
 
+## [0.12.1] -- 2025-08-26
+
+### Changed
+
+- More consistent documentation of parameters type.
+
+### Fixed
+
+- `cython` transfer matrices were not used
+
 ## [0.12.0] -- 2025-08-25
 
 ### Added